requirements.txt
bottle==0.13.4
cachetools==5.5.0
cbor2==5.6.5
certifi==2024.8.30
charset-normalizer==3.4.0
cheroot==10.0.1
exif==1.6.0
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gphoto2==2.6.2
idna==3.10
jaraco.functools==4.1.0
luma.core==2.4.2
luma.oled==3.14.0
more-itertools==10.5.0
mpu6050-raspberrypi==1.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
plum-py==0.8.7
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyftdi==0.55.4
pynmeagps==1.0.43
pyserial==3.5
pyusb==1.2.1
quat==1.0.0
requests==2.32.3
RPi.GPIO==0.7.1
rsa==4.9
smbus==1.1.post2
smbus2==0.5.0
spidev==3.6
urllib3==2.2.3
